Split EolDebugger into Eclipse-agnostic abstract class + Eclipse-specific subclass
