+++ plugins/org.eclipse.epsilon.ecl.dt/src/org/eclipse/epsilon/ecl/dt/launching/EclLaunchConfigurationDelegate.java
@@ -13,7 +13,7 @@ import org.eclipse.core.runtime.CoreException;
 import org.eclipse.debug.core.ILaunchConfiguration;
 import org.eclipse.epsilon.common.dt.launching.extensions.ModuleImplementationExtension;
 import org.eclipse.epsilon.ecl.IEclModule;
-import org.eclipse.epsilon.eol.dt.debug.EolDebugger;
+import org.eclipse.epsilon.eol.debug.EolDebugger;
 import org.eclipse.epsilon.eol.dt.launching.EpsilonLaunchConfigurationDelegate;
 
 public class EclLaunchConfigurationDelegate extends EpsilonLaunchConfigurationDelegate {
