+++ plugins/org.eclipse.epsilon.evl.dt/src/org/eclipse/epsilon/evl/dt/launching/EvlDebugger.java
@@ -10,7 +10,7 @@
 package org.eclipse.epsilon.evl.dt.launching;
 
 import org.eclipse.epsilon.common.module.ModuleElement;
-import org.eclipse.epsilon.eol.dt.debug.EolDebugger;
+import org.eclipse.epsilon.eol.debug.EolDebugger;
 import org.eclipse.epsilon.evl.dom.Constraint;
 import org.eclipse.epsilon.evl.dom.ConstraintContext;
 import org.eclipse.epsilon.evl.dom.Fix;
