+++ plugins/org.eclipse.epsilon.pinset.dt/src/org/eclipse/epsilon/pinset/dt/launching/PinsetDebugger.java
@@ -10,7 +10,7 @@
 package org.eclipse.epsilon.pinset.dt.launching;
 
 import org.eclipse.epsilon.common.module.ModuleElement;
-import org.eclipse.epsilon.eol.dt.debug.EolDebugger;
+import org.eclipse.epsilon.eol.debug.EolDebugger;
 import org.eclipse.epsilon.pinset.DatasetRule;
 
 /**
