+++ plugins/org.eclipse.epsilon.workflow/src/org/eclipse/epsilon/workflow/tasks/hosts/EclipseHost.java
@@ -30,9 +30,9 @@ import org.eclipse.epsilon.egl.dt.debug.EgxDebugger;
 import org.eclipse.epsilon.eml.IEmlModule;
 import org.eclipse.epsilon.eml.dt.launching.EmlDebugger;
 import org.eclipse.epsilon.eol.IEolModule;
+import org.eclipse.epsilon.eol.debug.EolDebugger;
 import org.eclipse.epsilon.eol.dt.ExtensionPointToolNativeTypeDelegate;
 import org.eclipse.epsilon.eol.dt.debug.EolDebugTarget;
-import org.eclipse.epsilon.eol.dt.debug.EolDebugger;
 import org.eclipse.epsilon.eol.dt.userinput.JFaceUserInput;
 import org.eclipse.epsilon.eol.models.IModel;
 import org.eclipse.epsilon.epl.IEplModule;
